[AWS Sagemaker] Move operations.py to train_utils.py
